js/xpconnect/tests/mochitest/test_bug500691.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=500691
-->
<head>
  <title>Test for Bug 500691</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=500691">Mozilla Bug 500691</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 500691 **/
ok(Function === alert.constructor, "alert's constructor is our Function");
ok(window.Function === alert.constructor, "window.Function is also correct");

</script>
</pre>
</body>
</html>
